requirements.txt
fastapi
uvicorn
python-dotenv
supabase
face_recognition
python-multipart
numpy
fpdf
requests
